requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.4
django-crispy-forms==2.1
django-tinymce==4.0.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
